Update the schedulers

+++ Core/Schedulers/scheduler.cpp
@@ -1,26 +1,26 @@
 #include "scheduler.h"
 
-unsigned long long Scheduler::getArrivalTimeSum()
+double Scheduler::getArrivalTimeSum()
 {
     return arrivalTimeSum;
 }
 
-unsigned long long Scheduler::getBurstTimeSum()
+double Scheduler::getBurstTimeSum()
 {
     return burstTimeSum;
 }
 
-unsigned long long Scheduler::getFinishTimeSum()
+double Scheduler::getFinishTimeSum()
 {
     return finishTimeSum;
 }
 
-unsigned long long Scheduler::getNumberOfProcesses()
+double Scheduler::getNumberOfProcesses()
 {
     return numberOfProcesses;
 }
 
 double Scheduler::getAverageWaitingTime()
 {
-    return (static_cast<double>(finishTimeSum - arrivalTimeSum - burstTimeSum) / numberOfProcesses);
+    return ((finishTimeSum - arrivalTimeSum - burstTimeSum) / numberOfProcesses);
 }
